+++ changelog.md
@@ -8,6 +8,8 @@ This version is not yet released. If you are reading this on the website, then t
 - Stabilize [`un °`](https://uiua.org/docs/un) [`under ⍜`](https://uiua.org/docs/under) for monadic functions
 - Add [`un °`](https://uiua.org/docs/un) [`group ⊕`](https://uiua.org/docs/group) and [`un °`](https://uiua.org/docs/un) [`partition ⊜`](https://uiua.org/docs/partition) for monadic functions
 - Stabilize [`negate ¯`](https://uiua.org/docs/negate) subscripts
+- Add experimental [lexical ordering](https://uiua.org/docs/experimental#lexical-ordering) syntax
+  - This enables some function packs and array notation to execute in the order they are normally read
 
 ## 0.15.0 - 2025-04-03
 You can find the release announcement [here](https://uiua.org/blog/uiua-0.15.0).